% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\docType{data}
\name{api_base}
\alias{api_base}
\title{PubChem API Base URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
api_base
}
\description{
The base URL for the PubChem API.
}
\keyword{datasets}
