% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{call_params}
\alias{call_params}
\alias{call_params.PubChemRequest}
\title{Call Function Parameters}
\usage{
call_params(object, ...)

\method{call_params}{PubChemRequest}(object)
}
\arguments{
\item{object}{<details here>}

\item{...}{<details here>}
}
\value{
a list
}
\description{
Call Function Parameters
}
\examples{
1L

}
