% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\docType{data}
\name{CompoundIdType}
\alias{CompoundIdType}
\title{Compound ID Types}
\format{
An object of class \code{list} of length 8.
}
\usage{
CompoundIdType
}
\description{
List of compound ID types.
}
\keyword{datasets}
