% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\docType{data}
\name{CoordinateType}
\alias{CoordinateType}
\title{Coordinate Types}
\format{
An object of class \code{list} of length 15.
}
\usage{
CoordinateType
}
\description{
List of coordinate types.
}
\keyword{datasets}
