% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\docType{data}
\name{ELEMENTS}
\alias{ELEMENTS}
\title{Elements}
\format{
An object of class \code{character} of length 118.
}
\usage{
ELEMENTS
}
\description{
Vector of chemical elements.
}
\keyword{datasets}
