% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_api.R
\name{pc_batch}
\alias{pc_batch}
\title{Batch-Orchestrate PubChem Workflows}
\usage{
pc_batch(ids, fn, chunk_size = 100, parallel = FALSE, workers = NULL, ...)
}
\arguments{
\item{ids}{Identifier vector.}

\item{fn}{Function to run on each chunk of `ids`.}

\item{chunk_size}{Chunk size.}

\item{parallel}{Logical; use parallel execution.}

\item{workers}{Number of workers.}

\item{...}{Additional arguments passed into `fn`.}
}
\value{
A typed `PubChemBatchResult` object.
}
\description{
Batch-Orchestrate PubChem Workflows
}
