% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_api.R
\name{pc_property}
\alias{pc_property}
\title{Query Compound Properties via the Next-Generation API}
\usage{
pc_property(
  identifier,
  properties,
  namespace = "cid",
  searchtype = NULL,
  options = NULL,
  ...
)
}
\arguments{
\item{identifier}{Identifier(s).}

\item{properties}{Character vector of property names.}

\item{namespace}{PubChem namespace.}

\item{searchtype}{Search type.}

\item{options}{Named list of query options.}

\item{...}{Additional arguments forwarded to `httr::RETRY`.}
}
\value{
A typed `PubChemRecord` object.
}
\description{
Query Compound Properties via the Next-Generation API
}
