% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_phase3.R
\name{pc_to_chemminer}
\alias{pc_to_chemminer}
\title{Convert PubChem Tables to ChemmineR SDF Objects}
\usage{
pc_to_chemminer(x, smiles_col = "CanonicalSMILES")
}
\arguments{
\item{x}{Input table or `PubChemResult`.}

\item{smiles_col}{Column with SMILES strings.}
}
\value{
A `ChemmineR` SDF object.
}
\description{
Converts SMILES strings from a PubChem table to a `ChemmineR` SDF object.
}
\details{
Requires the optional `ChemmineR` package with an available `smiles2sdf()`
function.
}
\examples{
names(formals(pc_to_chemminer))

\dontrun{
ex_tbl <- tibble::tibble(CanonicalSMILES = "CCO")
sdf <- pc_to_chemminer(ex_tbl)
class(sdf)
}
}
