% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_phase3.R
\name{pc_to_rcdk}
\alias{pc_to_rcdk}
\title{Convert PubChem Tables to rcdk Molecules}
\usage{
pc_to_rcdk(x, smiles_col = "CanonicalSMILES", id_col = "CID")
}
\arguments{
\item{x}{Input table or `PubChemResult`.}

\item{smiles_col}{Column with SMILES strings.}

\item{id_col}{Optional identifier column used to name returned molecules.}
}
\value{
A list of `rcdk` molecule objects.
}
\description{
Converts SMILES strings from a PubChem table into `rcdk` molecule objects.
}
\details{
Requires the optional `rcdk` package. Empty or missing SMILES values are
dropped before conversion.
}
\examples{
names(formals(pc_to_rcdk))

\dontrun{
ex_tbl <- tibble::tibble(CID = "1", CanonicalSMILES = "CCO")
mols <- pc_to_rcdk(ex_tbl)
length(mols)
}
}
