% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\docType{data}
\name{ProjectCategory}
\alias{ProjectCategory}
\title{Project Categories}
\format{
An object of class \code{list} of length 11.
}
\usage{
ProjectCategory
}
\description{
List of project categories.
}
\keyword{datasets}
