% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\name{property_map}
\alias{property_map}
\title{Property Map}
\usage{
property_map(
  x,
  type = c("match", "contain", "start", "end", "all"),
  .ignore.case = TRUE,
  ...
)
}
\description{
Map of properties to their respective names.
}
