% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_all_classes.R
\docType{class}
\name{PubChemInstance-class}
\alias{PubChemInstance-class}
\title{\code{PubChemInstanceList} and \code{PubChemInstance} objects}
\description{
\code{PubChemInstanceList} object is the super-class returned by a request of compound(s) from
PubChem Database, e.g., the output from \link{get_compounds}, \link{get_assays}, etc.
}
\note{
There is no constructor function for \code{'PubChemInstanceList'} or \code{'PubChemInstance'} classes. These object are
constructed within related functions and returned as output of PubChem requests.
}
\section{Slots}{

\describe{
  \item{\code{results}:}{a list with elements of each of requested compounds, assays, substances, etc.}
  \item{\code{request_args}:}{a list with function inputs of a PubChem request.}
  \item{\code{success}:}{a logical. Returns TRUE if a request is successfully completed.}
  \item{\code{error}:}{a list. If a request encountered an error, details are given within \code{error} slot.}
}
}

