% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Request Function for PubChem API}
\usage{
request(
  identifier = NULL,
  namespace = "cid",
  domain = "compound",
  operation = NULL,
  output = "JSON",
  searchtype = NULL,
  ...
)
}
\arguments{
\item{identifier}{The identifier for the compound.}

\item{namespace}{The namespace for the identifier (default: 'cid').}

\item{domain}{The domain for the request (default: 'compound').}

\item{operation}{The operation to be performed (default: NULL).}

\item{output}{The desired output format (default: 'JSON').}

\item{searchtype}{The type of search to be performed (default: NULL).}

\item{...}{Additional parameters. Currently has no effect on the results.}
}
\value{
A constructed URL for the PubChem API.
}
\description{
Constructs a URL for the PubChem API based on the provided parameters.
}
\examples{
request(
  identifier = "aspirin",
  namespace = "name"
)
}
