% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{retrieve}
\alias{retrieve}
\alias{retrieve.PubChemInstance}
\alias{retrieve.PubChemInstanceList}
\alias{retrieve.PC_Substance}
\title{Retrieve Information from PubChem Instances}
\usage{
retrieve(object, ...)

\method{retrieve}{PubChemInstance}(object, .slot = NULL, .to.data.frame = TRUE, .verbose = FALSE, ...)

\method{retrieve}{PubChemInstanceList}(
  object,
  .which = NULL,
  .slot = NULL,
  .to.data.frame = TRUE,
  .combine.all = FALSE,
  ...
)

\method{retrieve}{PC_Substance}(
  object,
  .slot = NULL,
  .idx = 1,
  .to.data.frame = TRUE,
  .verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{An object returned from a PubChem request.}

\item{...}{Additional arguments passed to other methods. Currently, these have no effect.}

\item{.slot}{A string specifying which slot to return. Should not be NULL with some exceptions. See notes for details.}

\item{.to.data.frame}{a logical. If TRUE, returned object will be forced to be converted into a data.frame (or tibble).
If failed to convert into a data.frame, a list will be returned with a warning. Be careful for complicated lists
(i.e., many elements nested within each other) since it may be time consuming to convert such lists into a data frame.
Furthermore, \code{.to.data.frame} is ignored for specific scenarios.}

\item{.verbose}{a logical. Should the resulting object printed into R Console? If TRUE the object is returned invisibly
and the output is nicely printed to R Console. This option may not be available for some slots (or classes).
See Notes/Details.}

\item{.combine.all}{a logical. If TRUE, properties of all requested instances are combined into single
data frame (or list if \code{.to.data.frame = FALSE}).}
}
\description{
This generic function extracts a specific slot from a PubChem instance.
}
\note{
If object is from \code{'PC_Properties'} class, \code{.slot} definition will be omitted and all the requested
properties will be retrieved from \code{object}.

\subsection{Use of \code{'.verbose'} argument}
\code{retrieve} returns output silently (invisible) when \code{.verbose = TRUE}. However, the function treats differently
under following scenarios:
\itemize{
  \item{\code{.verbose} is ignored if \code{.combine.all = TRUE}. Output is returned silently.}
  \item{\code{.verbose} is ignored if requested slot is not printable to R Console because it is too complicated to print.}
}
}
\examples{
compounds <- get_compounds(identifier = c(
  "aspirin", "ibuprofen", "rstudio"),
  namespace = "name"
 )

# Extract information for "aspirin"
aspirin <- instance(compounds, "aspirin")
# print(aspirin)

# Extract a specific slot from "aspirin" compound.
retrieve(aspirin, "props", .to.data.frame = TRUE)

# EXAMPLES (PubChemInstanceList)
retrieve(compounds, "aspirin", "props", .to.data.frame = TRUE)

# Verbose Assay References to R Console
assays <- get_assays(identifier = c(1234, 7815), namespace = "aid")

instance(assays, "7815")
retrieve(assays, "7815", "xref", .verbose = TRUE)

# Print assay protocol to R Console (if available)
# Note that it may be too long to print for some assays.
# retrieve(assays, "1234", "protocol", .verbose = TRUE)

# No protocol is available for assay "1234".
# retrieve(assays, "7815", "protocol", .verbose = TRUE)

# Ignores ".verbose" and ".which" if ".combine.all = TRUE".
retrieve(assays, .slot = "xref", .verbose = TRUE, .combine.all = TRUE)

}
