% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_sources.R
\name{get_all_sources}
\alias{get_all_sources}
\title{Retrieve All Sources from PubChem}
\usage{
get_all_sources(domain = "substance")
}
\arguments{
\item{domain}{A character string specifying the domain for which sources ('substance', 'assay') are to be retrieved.
Default is 'substance'.}
}
\value{
A character vector containing the names of all sources for the specified domain.
}
\description{
This function retrieves a list of all current depositors of substances or assays from PubChem.
}
\examples{
get_all_sources(
  domain = 'substance'
)
}
