% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_api.R
\name{pc_cache_clear}
\alias{pc_cache_clear}
\title{Clear PubChemR Request Cache}
\usage{
pc_cache_clear(cache_dir = NULL, memory = TRUE, disk = TRUE)
}
\arguments{
\item{cache_dir}{Cache directory. If `NULL`, uses configured cache directory.}

\item{memory}{Logical; clear in-memory cache.}

\item{disk}{Logical; clear on-disk cache.}
}
\value{
Invisibly returns `TRUE`.
}
\description{
Clear PubChemR Request Cache
}
