% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_api.R
\name{pc_collect}
\alias{pc_collect}
\title{Collect Results From an Async PubChem Query}
\usage{
pc_collect(x, ...)
}
\arguments{
\item{x}{A `PubChemAsyncQuery` object.}

\item{...}{Additional arguments passed to `pc_poll`.}
}
\value{
A `PubChemResult` object.
}
\description{
Collect Results From an Async PubChem Query
}
