% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_api.R
\name{pc_submit}
\alias{pc_submit}
\title{Submit an Asynchronous PubChem Query}
\usage{
pc_submit(
  domain = "compound",
  namespace = "cid",
  identifier = NULL,
  operation = NULL,
  searchtype = NULL,
  output = "JSON",
  options = NULL,
  ...
)
}
\arguments{
\item{domain}{PubChem domain.}

\item{namespace}{PubChem namespace.}

\item{identifier}{Identifier(s).}

\item{operation}{Operation.}

\item{searchtype}{Search type.}

\item{output}{Output format.}

\item{options}{Named list of query options.}

\item{...}{Additional arguments forwarded to `httr::RETRY`.}
}
\value{
An object of class `PubChemAsyncQuery`.
}
\description{
Submit an Asynchronous PubChem Query
}
