% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_api.R
\name{pc_substance}
\alias{pc_substance}
\title{Query Substance Records via the Next-Generation API}
\usage{
pc_substance(
  identifier,
  namespace = "sid",
  operation = "record",
  output = "JSON",
  options = NULL,
  ...
)
}
\arguments{
\item{identifier}{Identifier(s).}

\item{namespace}{PubChem namespace.}

\item{operation}{Operation.}

\item{output}{Output format.}

\item{options}{Named list of query options.}

\item{...}{Additional arguments forwarded to `httr::RETRY`.}
}
\value{
A typed `PubChemRecord` object.
}
\description{
Query Substance Records via the Next-Generation API
}
