% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\docType{data}
\name{property_map}
\alias{property_map}
\title{Property Map}
\format{
An object of class \code{list} of length 40.
}
\usage{
property_map
}
\description{
Map of properties to their respective names.
}
\keyword{datasets}
