% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{pubChemData}
\alias{pubChemData}
\alias{pubChemData.PugRestInstance}
\title{Retrieve Raw Data from PUG REST Object}
\usage{
pubChemData(object, ...)

\method{pubChemData}{PugRestInstance}(object, ...)
}
\arguments{
\item{object}{an object of class 'PugRestInstance' returned from \link{get_pug_rest} function.}

\item{...}{additional arguments. Currently has no effect on results.}
}
\value{
a vector, list, or data.frame containing the raw data retrieved from Pub Chem database through PUG REST API.
}
\description{
A short description...
}
\examples{
\donttest{
result <- get_pug_rest(identifier = "2244", namespace = "cid", domain = "compound", output = "JSON")
pubChemData(result)
}

}
\seealso{
\link{get_pug_rest}
}
