% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{request_args}
\alias{request_args}
\alias{request_args.PubChemInstance}
\alias{request_args.PubChemInstance_AIDs}
\title{Call Function Parameters}
\usage{
request_args(object, ...)

\method{request_args}{PubChemInstance}(object)

\method{request_args}{PubChemInstance_AIDs}(object)
}
\arguments{
\item{object}{<details here>}

\item{...}{<details here>}
}
\value{
a list
}
\description{
Call Function Parameters
}
\examples{
1L

}
