% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{request_args}
\alias{request_args}
\title{Call Function Parameters}
\usage{
request_args(object, .which = NULL, ...)
}
\arguments{
\item{object}{an object of class}

\item{...}{<details here>}
}
\value{
a list
}
\description{
Call Function Parameters
}
\examples{
1L

}
