% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{synonyms}
\alias{synonyms}
\alias{synonyms.PubChemInstance_Synonyms}
\title{Getter function for 'Synonyms'}
\usage{
synonyms(object, ...)

\method{synonyms}{PubChemInstance_Synonyms}(object, .to.data.frame = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{'PubChemInstance_Synonyms'}.}

\item{...}{Additional arguments passed to other methods. Currently, these have no effect.}

\item{.to.data.frame}{a logical. If TRUE, returned object will be forced to be converted into a data.frame (or tibble).
If failed to convert into a data.frame, a list will be returned with a warning. Be careful for complicated lists
(i.e., many elements nested within each other) since it may be time consuming to convert such lists into a data frame.}
}
\value{
A \code{data.frame} (or \code{list}) object containing the synonym data.
}
\description{
Extracts synonym data from a PubChem request using the function \link{get_synonyms}.
}
\examples{
\donttest{
syns <- get_synonyms(identifier = c("aspirin", "caffeine"), namespace = "name")
synonyms(syns)
}

}
